+++ backend/src/main/java/com/myong/backend/controller/DesignerController.java
@@ -3,13 +3,19 @@ package com.myong.backend.controller;
 
 import com.myong.backend.domain.dto.designer.Api;
 import com.myong.backend.domain.dto.designer.SignUpRequest;
+import com.myong.backend.domain.dto.email.EmailCheckDto;
+import com.myong.backend.domain.dto.email.EmailRequestDto;
 import com.myong.backend.service.DesignerService;
+import com.myong.backend.service.EmailSendService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.HashMap;
+import java.util.Map;
 
 
 @Slf4j
@@ -55,5 +61,32 @@ public class DesignerController {
         log.info("checked nickname duplicate: {}", nickname);
         //중복되면 true, 중복안되면 false
         return ResponseEntity.ok(designerService.checkNicknameDuplication(nickname));
+    }
+
+    private final EmailSendService emailSendService;
+
+    //Send Email: 이메일 전송 버튼 클릭시
+    @PostMapping("/sendemail")
+    public Map<String, String> mailSend(
+            @RequestBody EmailRequestDto emailRequestDto
+    ){
+        String code = emailSendService.joinEmail(emailRequestDto.getEmail());
+        //response를 Json으로 변환
+        Map<String, String> response = new HashMap<>();
+        response.put("code", code);
+
+        return response;
+    }
+
+    //이메일 인증
+
+    @PostMapping("/verifyemail")
+    public String authCheck(@RequestBody @Valid EmailCheckDto emailCheckDto){
+        Boolean checked = emailSendService.checkAuthNum(emailCheckDto.getEmail(), emailCheckDto.getAuthNum());
+        if(checked){
+            return "이메일 인증 성공!!";
+        }else {
+            throw new NullPointerException("이메일 인증 실패");
+        }
     }
 }
